requirements.txt
python-telegram-bot[ext]
sqlalchemy[asyncio]
psycopg2-binary # Or your preferred database driver
pandas
matplotlib
dataframe_image
numpy
apscheduler
python-dotenv